Add new Component for booklist
